+++ src/main/java/app/model/MultiplayerModel.java
@@ -27,19 +27,19 @@ public class MultiplayerModel
 
   /** Update interval in sec. */
   private static final long POLLING_INTERVAL = 1;
-
   private static final long FALL_BACK_START_DELAY = 3;
+  private static final FinishedMessage FINISHED = new FinishedMessage("FINISHED","waiting for race end");
+  private static final FinishedMessage NOT_FINISHED = new FinishedMessage("HURRY!", "");
 
   private MultiplayerModelObserver observer;
 
   private long raceStart;
-  private long raceEnd;
   private final Typeracer typeracer;
   private final RaceData raceData;
   private List<PlayerUpdate> updates;
   private State state;
   private final ScheduledExecutorService scheduler;
-  private String hurry;
+  private FinishedMessage finishedMessage;
 
   private enum State {
     PRE_START,
@@ -93,11 +93,10 @@ public class MultiplayerModel
     CheckResult check = typeracer.check(key.charAt(0));
     if (typeracer.getState().getCurrentGamePhase() == GamePhase.FINISHED) {
       sendProgress();
-      this.hurry = "FINISHED!";
+      this.finishedMessage = FINISHED;
       this.state = State.CHECKERED_FLAG;
-      this.raceEnd = Timestamp.currentTimestamp();
     } else {
-      this.hurry = "HURRY!";
+      this.finishedMessage = NOT_FINISHED;
     }
     return check;
   }
@@ -163,11 +162,9 @@ public class MultiplayerModel
   @Override
   public void receivedCheckeredFlag(long raceStop) {
     this.state = State.CHECKERED_FLAG;
-    this.raceEnd = raceStop;
     if (observer != null) {
       Platform.runLater(
           () -> {
-            observer.updatedCountDown(this.raceEnd - Timestamp.currentTimestamp());
             observer.checkeredFlag(raceStop);
           });
     }
@@ -236,5 +233,5 @@ public class MultiplayerModel
     return typeracer.getState().getTypeChar().getCounter();
   }
 
-  public String getHurryUp() { return this.hurry; };
+  public FinishedMessage getFinishedText() { return this.finishedMessage; };
 }