+++ src/main/java/protocol/Request.java
@@ -16,6 +16,7 @@ public class Request {
     public static final String LOBBY_UPDATE = "get lobby update";
     public static final String PREV_RACE_RESULT = "previous race result";
     public static final String CHAT_MESSAGE = "chat message";
+    public static final String CHAT_HISTORY = "chat history";
   }
 
   public final String type;